Cancel Feature Added
Since customers may select XRP and not pay, we don't want to have pending orders just sitting in the Shopify queue. We entered a cancel time field on the menu and after, for example, 15 minutes has passed since the order was entered, the program will automatically cancel the order and send notification to the customer that the order has been canceled.